# docs/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: ModelMuxer API
  description: |
    ModelMuxer is an intelligent LLM routing engine that provides cost-optimized, 
    high-quality AI model access through advanced routing strategies including 
    cascade routing, semantic routing, and heuristic routing.
    
    ## Features
    - **Multi-Provider Support**: OpenAI, Anthropic, Google, Mistral, Groq, and more
    - **Intelligent Routing**: Cascade, semantic, heuristic, and hybrid routing
    - **Cost Optimization**: Budget management and cost-aware routing
    - **Enterprise Features**: Multi-tenancy, RBAC, audit logging
    - **Production Ready**: Monitoring, security, and compliance features
    
    ## Authentication
    ModelMuxer supports multiple authentication methods:
    - **Bearer Token**: JWT tokens for user authentication
    - **API Key**: Organization-scoped API keys for programmatic access
    
    ## Rate Limiting
    API endpoints are rate-limited based on your subscription plan:
    - **Free**: 1,000 requests/month
    - **Starter**: 10,000 requests/month  
    - **Professional**: 100,000 requests/month
    - **Enterprise**: Custom limits
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information
    in JSON format with error codes, messages, and additional context.
  version: 1.0.0
  contact:
    name: ModelMuxer Support
    url: https://docs.modelmuxer.com
    email: support@modelmuxer.com
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /v1/chat/completions:
    post:
      summary: Create Chat Completion
      description: |
        Create a chat completion using ModelMuxer's intelligent routing.
        The system will automatically select the best model based on your
        routing strategy and requirements.
      tags: [Chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              simple_request:
                summary: Simple chat completion
                value:
                  messages:
                    - role: user
                      content: "What is the capital of France?"
                  temperature: 0.7
                  max_tokens: 100
              cascade_request:
                summary: Cascade routing request
                value:
                  messages:
                    - role: user
                      content: "Explain quantum computing in simple terms"
                  routing_strategy: "cascade"
                  cascade_config:
                    cascade_type: "quality_focused"
                    max_budget: 0.05
                    quality_threshold: 0.8
      responses:
        '200':
          description: Successful completion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication endpoint
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Organization API key for programmatic access
  schemas:
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
          description: The role of the message author
        content:
          type: string
          description: The content of the message
        name:
          type: string
          description: Optional name of the message author
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          description: List of messages in the conversation
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          description: Sampling temperature for randomness
        max_tokens:
          type: integer
          minimum: 1
          maximum: 4096
          description: Maximum number of tokens to generate
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
          description: Nucleus sampling parameter
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
          description: Frequency penalty for repetition
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
          description: Presence penalty for new topics
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: Stop sequences for generation
        stream:
          type: boolean
          default: false
          description: Enable streaming responses
...
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the completion
        object:
          type: string
          example: "chat.completion"
        created:
          type: integer
          description: Unix timestamp of creation
        model:
          type: string
          description: Model used for the completion
        choices:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              message:
                $ref: '#/components/schemas/ChatMessage'
              finish_reason:
                type: string
                enum: [stop, length, content_filter, null]
        usage:
          type: object
          properties:
            prompt_tokens:
              type: integer
            completion_tokens:
              type: integer
            total_tokens:
              type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Additional error details
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid request parameters"
            code: "INVALID_REQUEST"
            timestamp: "2025-01-15T10:30:00Z"
    Unauthorized:
      description: Unauthorized - invalid or missing authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid or missing authentication token"
            code: "UNAUTHORIZED"
            timestamp: "2025-01-15T10:30:00Z"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded: 100/60 requests per minute"
            code: "RATE_LIMIT_EXCEEDED"
            timestamp: "2025-01-15T10:30:00Z"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2025-01-15T10:30:00Z"
...